Refactor motif page and genomicSites
